requirements.txt
slack-bolt==1.21.3
fastapi==0.115.6
uvicorn[standard]==0.34.0
pyairtable==2.3.3
anthropic==0.43.0
python-dotenv==1.0.1
httpx==0.28.1
